Move core/python2 to build dep.
Because we're using a virtualenv we only need the venv's own Python runtime.
Thsi doesn't do anything for the resulting size of the Habitat native package
but it does make starting the package quicker as there are less dependencies
to fetch.  For Docker containers it yields about a 100MB reduction in
container size.